tut3/requirements.txt
absl-py==0.12.0
alabaster==0.7.12
alembic==1.5.8
amqp==5.0.5
anyjson==0.3.3
apipkg==1.5
appdirs==1.4.4
argon2-cffi==20.1.0
Arpeggio==1.10.1
asgiref==3.3.1
asn1crypto==1.4.0
astroid==2.5.1
astunparse==1.6.3
async-generator==1.10
atomicwrites==1.3.0
attrs==20.3.0
Automat==20.2.0
awscli==1.19.22
Babel==2.8.0
backcall==0.1.0
backports-abc==0.5
backports.functools-lru-cache==1.6.1
bcrypt==3.1.7
beautifulsoup4==4.9.0
billiard==3.6.2.0
bleach==3.3.0
blessings==1.7
blurb==1.0.8
bokeh==2.3.0
boto==2.49.0
boto3==1.17.22
botocore==1.20.22
cachetools==4.2.1
celery==5.0.0
certifi==2022.9.14
cffi==1.14.3
chardet==3.0.4
charset-normalizer==2.1.1
click==7.1.2
client==0.0.1
cliff==3.7.0
cloudpickle==1.6.0
cmake==3.17.3
cmd2==1.5.0
codecov==2.0.22
colorama==0.4.3
configobj==5.0.6
constantly==15.1.0
coverage==5.5
coveralls==3.0.1
cppy==1.1.0
crcmod==1.7
cryptography==3.3.2
cycler==0.10.0
Cython==0.29.22
cytoolz==0.9.0.1
dask==2021.3.1
datashape==0.5.2
debtcollector==1.21.0
decorator==4.4.2
defusedxml==0.7.1
distlib==0.3.1
Django==3.1.7
django-debug-toolbar==2.2
django-extensions==3.1.1
djangorestframework==3.12.4
dnspython==1.16.0
doc8==0.8.0
docker-py==1.10.6
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
dulwich==0.20.21
ecdsa==0.16.1
elasticsearch==7.11.0
entrypoints==0.3
eventlet==0.30.2
execnet==1.8.0
extras==1.0.0
filelock==3.0.12
fixtures==3.0.0
flake8==3.9.0
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==20210329160745
flit==2.1.0
flit-core==2.3.0
fsspec==0.8.7
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
gevent==1.4.0
google-api-core==1.23.0
google-api-python-client==2.0.2
google-auth==1.27.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
google-cloud==0.34.0
google-cloud-bigquery==2.11.0
google-cloud-bigquery-datatransfer==3.1.0
google-cloud-bigtable==1.7.0
google-cloud-container==2.4.0
google-cloud-core==1.6.0
google-cloud-datastore==2.0.1
google-cloud-dns==0.32.2
google-cloud-error-reporting==1.1.1
google-cloud-firestore==2.0.1
google-cloud-iam==0.1.0
google-cloud-language==1.1.1
google-cloud-logging==2.1.1
google-cloud-monitoring==2.2.0
google-cloud-pubsub==0.39.1
google-cloud-resource-manager==0.30.3
google-cloud-runtimeconfig==0.32.1
google-cloud-spanner==3.3.0
google-cloud-speech==2.2.0
google-cloud-storage==1.36.1
google-cloud-trace==0.20.2
google-cloud-translate==3.0.2
google-cloud-videointelligence==1.16.1
google-cloud-vision==2.2.0
google-crc32c==1.1.2
google-pasta==0.2.0
google-resumable-media==1.2.0
googleapis-common-protos==1.53.0
greenlet==1.0.0
grpc-google-iam-v1==0.12.3
grpcio==1.33.2
gunicorn==20.1.0
h11==0.13.0
h5py==2.10.0
hdfs3==0.3.0
html5lib==1.0.1
httplib2==0.19.0
hyperlink==21.0.0
hypothesis==6.6.0
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==3.10.0
importlib-resources==5.1.2
incremental==21.3.0
iniconfig==1.1.1
intreehooks==1.0
invoke==1.5.0
ipaddress==1.0.23.1
ipykernel==5.5.2
ipython==7.22.0
ipython-genutils==0.2.0
ipywidgets==7.6.0
iso8601==0.1.14
isodate==0.6.0
isort==4.3.21
itsdangerous==1.1.0
jedi==0.18.0
jeepney==0.6.0
Jinja2==2.11.3
jmespath==0.10.0
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyter-packaging==0.7.12
jupyterlab-widgets==1.0.0
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==23.0.1
keystoneauth1==4.3.1
kiwisolver==1.2.0
kombu==5.0.2
lazy-object-proxy==1.4.3
libcst==0.3.17
linecache2==1.0.0
locket==0.2.1
lockfile==0.12.2
logilab-common==1.8.0
luigi==2.8.3
lxml==4.6.2
m2r==0.2.1
Mako==1.1.4
Markdown==3.3.4
MarkupSafe==1.1.1
matplotlib==3.3.0
mccabe==0.6.1
meld3==1.0.2
mistune==0.8.4
mock==4.0.2
monotonic==1.5
more-itertools==8.6.0
mpmath==1.2.1
msgpack==1.0.1
multipledispatch==0.6.0
mypy==0.812
mypy-extensions==0.4.3
nbconvert==5.6.1
nbformat==5.1.2
netaddr==0.7.19
netifaces==0.10.9
networkx==2.4
nltk==3.5
nose==1.3.7
notebook==6.3.0
numexpr==2.7.3
numpy==1.19.4
oauth2client==4.1.3
oauthlib==3.1.0
openstackdocstheme==1.29.2
opt-einsum==3.3.0
os-service-types==1.6.0
oslo.config==8.5.0
oslo.i18n==3.23.1
oslo.serialization==4.1.0
oslo.utils==4.8.0
outcome==1.2.0
packaging==20.9
pandas==1.2.3
pandocfilters==1.4.2
parso==0.8.1
partd==1.1.0
parver==0.2.1
passlib==1.7.2
Paste==3.5.0
PasteDeploy==2.1.0
pathlib2==2.3.5
pbr==5.4.5
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
pika==1.2.0
Pillow==8.1.2
pipenv==2018.11.26
pkgconfig==1.4.0
pkginfo==1.7.0
pluggy==0.13.1
ply==3.11
prettytable==2.1.0
prometheus-client==0.9.0
prompt-toolkit==3.0.16
proto-plus==1.14.2
protobuf==3.15.6
psutil==5.7.3
ptyprocess==0.6.0
py==1.10.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybind11==2.6.0
pycodestyle==2.7.0
pycparser==2.20
pycryptodome==3.10.1
pydal==20210215.1
pyflakes==2.3.0
Pygments==2.8.1
pylint==2.7.2
pymongo==3.11.3
PyMySQL==0.9.3
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyperclip==1.8.2
pyrsistent==0.15.4
PySocks==1.7.1
pystache==0.5.4
pytest==6.1.2
pytest-cov==2.11.1
pytest-forked==1.3.0
pytest-mock==3.3.1
pytest-runner==5.2
pytest-timeout==1.4.2
pytest-xdist==2.1.0
python-dateutil==2.8.1
python-dotenv==0.21.0
python-editor==1.0.4
python-gflags==3.1.2
python-keystoneclient==3.19.0
python-memcached==1.59
python-mimeparse==1.6.0
python-novaclient==17.4.0
python-pcre==0.7
python-swiftclient==3.11.1
pytoml==0.1.21
pytz==2020.4
PyWavelets==1.1.1
PyYAML==5.4.1
pyzmq==19.0.1
qtconsole==4.7.7
raven==6.10.0
readme-renderer==29.0
redis==3.5.3
regex==2021.3.17
reno==2.9.2
repoze.lru==0.7
requests==2.6.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
restructuredtext-lint==1.3.1
retrying==1.3.3
rfc3986==1.4.0
rsa==4.7.2
s3transfer==0.3.3
scandir==1.10.0
scikit-build==0.8.1
scikit-image==0.17.2
scipy==1.6.1
seaborn==0.10.1
SecretStorage==3.3.1
selenium==4.4.3
Send2Trash==1.5.0
service-identity==18.1.0
setuptools-git==1.2
setuptools-scm==6.0.1
simplegeneric==0.8.1
simplejson==3.17.2
six==1.15.0
sniffio==1.3.0
snowballstemmer==2.1.0
sortedcontainers==2.2.2
soupsieve==2.2.1
South==1.0.2
Sphinx==3.5.3
sphinx-rtd-theme==0.4.3
sphinxcontrib-apidoc==0.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-asyncio==0.2.0
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sphinxcontrib-websupport==1.2.2
SQLAlchemy==1.4.3
sqlparse==0.3.1
statsd==3.3.0
stevedore==3.3.0
sympy==1.7.1
tables==3.6.1
tensorboard==2.4.0
tensorflow==2.4.0
tensorflow-estimator==2.4.0
termcolor==1.1.0
terminado==0.9.4
testpath==0.4.2
testtools==2.3.0
thrift==0.13.0
thriftpy==0.3.9
toml==0.10.2
toolz==0.11.1
tornado==6.1
tox==3.23.0
tqdm==4.59.0
traceback2==1.4.0
traitlets==5.0.4
trio==0.21.0
trio-websocket==0.9.2
twine==3.4.1
Twisted==21.2.0
typed-ast==1.4.2
typing==3.7.4.3
typing-extensions==3.7.4.3
typing-inspect==0.6.0
ujson==4.0.2
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.26.4
vcversioner==2.16.0.0
vine==5.0.0
virtualenv==20.4.2
virtualenv-clone==0.5.1
waitress==2.0.0
wcwidth==0.1.9
webdriver-manager==3.8.3
webencodings==0.5.1
WebOb==1.8.7
websocket-client==0.58.0
WebTest==2.0.35
Werkzeug==1.0.1
whereto==0.4.0
widgetsnbextension==3.5.1
wrapt==1.12.1
wsproto==1.2.0
xlrd==2.0.1
zipp==3.4.1
zope.interface==5.1.0
